+++ apps/explorer/lib/explorer/chain.ex
@@ -1822,21 +1822,23 @@ defmodule Explorer.Chain do
   Lists the top `t:Explorer.Chain.Token.t/0`'s'.
 
   """
-  @spec list_top_tokens :: [{Token.t(), non_neg_integer()}]
-  def list_top_tokens(options \\ []) do
+  @spec list_top_tokens(String.t()) :: [{Token.t(), non_neg_integer()}]
+  def list_top_tokens(filter, options \\ []) do
     paging_options = Keyword.get(options, :paging_options, @default_paging_options)
 
-    fetch_top_tokens(paging_options)
+    fetch_top_tokens(filter, paging_options)
   end
 
-  @spec list_top_bridged_tokens :: [{Token.t(), non_neg_integer()}]
-  def list_top_bridged_tokens(options \\ []) do
+  @spec list_top_bridged_tokens(atom(), String.t(), [paging_options | necessity_by_association_option]) :: [
+          {Token.t(), non_neg_integer()}
+        ]
+  def list_top_bridged_tokens(destination, filter, options \\ []) do
     paging_options = Keyword.get(options, :paging_options, @default_paging_options)
 
-    fetch_top_bridged_tokens(paging_options)
+    fetch_top_bridged_tokens(destination, paging_options, filter)
   end
 
-  defp fetch_top_tokens(paging_options) do
+  defp fetch_top_tokens(filter, paging_options) do
     base_query =
       from(t in Token,
         where: t.total_supply > ^0,
@@ -1844,21 +1846,35 @@ defmodule Explorer.Chain do
         preload: [:contract_address]
       )
 
-    base_query
-    |> page_tokens(paging_options)
-    |> limit(^paging_options.page_size)
+    base_query_with_paging =
+      base_query
+      |> page_tokens(paging_options)
+      |> limit(^paging_options.page_size)
+
+    query =
+      if filter && filter !== "" do
+        base_query_with_paging
+        |> where(fragment("to_tsvector('english', symbol || ' ' || name ) @@ to_tsquery(?)", ^filter))
+      else
+        base_query_with_paging
+      end
+
+    query
     |> Repo.all()
   end
 
-  defp fetch_top_bridged_tokens(paging_options) do
+  defp fetch_top_bridged_tokens(destination, paging_options, filter) do
+    chain_id = translate_destination_to_chain_id(destination)
+
     bridged_tokens_query =
       from(bt in BridgedToken,
-        select: bt
+        select: bt,
+        where: bt.foreign_chain_id == ^chain_id
       )
 
     base_query =
       from(t in Token,
-        left_join: bt in subquery(bridged_tokens_query),
+        right_join: bt in subquery(bridged_tokens_query),
         on: t.contract_address_hash == bt.home_token_contract_address_hash,
         where: t.total_supply > ^0,
         where: t.bridged,
@@ -1867,10 +1883,29 @@ defmodule Explorer.Chain do
         preload: [:contract_address]
       )
 
-    base_query
-    |> page_tokens(paging_options)
-    |> limit(^paging_options.page_size)
+    base_query_with_paging =
+      base_query
+      |> page_tokens(paging_options)
+      |> limit(^paging_options.page_size)
+
+    query =
+      if filter && filter !== "" do
+        base_query_with_paging
+        |> where(fragment("to_tsvector('english', symbol || ' ' || name ) @@ to_tsquery(?)", ^filter))
+      else
+        base_query_with_paging
+      end
+
+    query
     |> Repo.all()
+  end
+
+  defp translate_destination_to_chain_id(destination) do
+    case destination do
+      :eth -> 1
+      :bsc -> 56
+      _ -> 1
+    end
   end
 
   @doc """
@@ -3843,7 +3878,17 @@ defmodule Explorer.Chain do
           set_token_bridged_status(token_address_hash, false)
 
         created_from_int_tx && created_from_int_tx_success ->
-          extract_omni_bridged_token_metadata_wrapper(token_address_hash, created_from_int_tx_success)
+          extract_omni_bridged_token_metadata_wrapper(
+            token_address_hash,
+            created_from_int_tx_success,
+            :eth_omni_bridge_mediator
+          )
+
+          extract_omni_bridged_token_metadata_wrapper(
+            token_address_hash,
+            created_from_int_tx_success,
+            :bsc_omni_bridge_mediator
+          )
 
         true ->
           :ok
@@ -3853,8 +3898,8 @@ defmodule Explorer.Chain do
     :ok
   end
 
-  defp extract_omni_bridged_token_metadata_wrapper(token_address_hash, created_from_int_tx_success) do
-    omni_bridge_mediator = Application.get_env(:block_scout_web, :omni_bridge_mediator)
+  defp extract_omni_bridged_token_metadata_wrapper(token_address_hash, created_from_int_tx_success, mediator) do
+    omni_bridge_mediator = Application.get_env(:block_scout_web, mediator)
     %{transaction_hash: transaction_hash} = created_from_int_tx_success
 
     if omni_bridge_mediator && omni_bridge_mediator !== "" do
@@ -5745,5 +5790,43 @@ defmodule Explorer.Chain do
     |> select([b], b.timestamp)
     |> limit(1)
     |> Repo.one()
+  end
+
+  def bridged_tokens_enabled? do
+    eth_omni_bridge_mediator = Application.get_env(:block_scout_web, :eth_omni_bridge_mediator)
+    bsc_omni_bridge_mediator = Application.get_env(:block_scout_web, :bsc_omni_bridge_mediator)
+
+    (eth_omni_bridge_mediator && eth_omni_bridge_mediator !== "") ||
+      (bsc_omni_bridge_mediator && bsc_omni_bridge_mediator !== "")
+  end
+
+  def bridged_tokens_eth_enabled? do
+    eth_omni_bridge_mediator = Application.get_env(:block_scout_web, :eth_omni_bridge_mediator)
+
+    eth_omni_bridge_mediator && eth_omni_bridge_mediator !== ""
+  end
+
+  def bridged_tokens_bsc_enabled? do
+    bsc_omni_bridge_mediator = Application.get_env(:block_scout_web, :bsc_omni_bridge_mediator)
+
+    bsc_omni_bridge_mediator && bsc_omni_bridge_mediator !== ""
+  end
+
+  def chain_id_display_name(nil), do: ""
+
+  def chain_id_display_name(chain_id) do
+    chain_id_int =
+      if is_integer(chain_id) do
+        chain_id
+      else
+        chain_id
+        |> Decimal.to_integer()
+      end
+
+    case chain_id_int do
+      1 -> "eth"
+      56 -> "bsc"
+      _ -> ""
+    end
   end
 end